+++ app/src/main/java/com/lostf1sh/pixelplayeross/di/Qualifiers.kt
@@ -29,3 +29,10 @@ annotation class BackupGson
 @Qualifier
 @Retention(AnnotationRetention.BINARY)
 annotation class AppScope
+
+/**
+ * Qualifier for ListenBrainz Retrofit instance.
+ */
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class ListenBrainzRetrofit